Timeline Tab to  multiple files.
